+++ common/src/main/java/dev/minhnh/yetanotherthirst/Constants.java
@@ -11,6 +11,7 @@ public class Constants {
 	public static final String MOD_ID = "yet_another_thirst";
 	public static final String MOD_NAME = "Yet Another Thirst";
 	public static final Logger LOG = LoggerFactory.getLogger(MOD_NAME);
+	public static final ResourceLocation THIRST_ICONS = asResource("textures/gui/thirst_icons.png");
 
 	@Nonnull
 	public static ResourceLocation asResource(@Nonnull String path) {